+++ lib/index.ts
@@ -1,4 +1,4 @@
-import { Inertia, VisitOptions } from '@inertiajs/inertia';
+import { Inertia } from '@inertiajs/inertia';
 
 import extendInertiaEventCallbacks from './events';
 
@@ -8,18 +8,13 @@ import { FormKitInertiaContext } from './types';
 export default (node: FormKitNode) => {
   if (node.props.type !== 'form' || !node.context) return;
 
-  const mergeOptions = (options?: VisitOptions) => ({
-    ...options,
-    ...extendInertiaEventCallbacks(node, options)
-  });
-
   (node.context.inertia as FormKitInertiaContext) = {
-    visit: (href, options?) => Inertia.visit(href, mergeOptions(options)),
-    reload: (options?) => Inertia.reload(mergeOptions(options)),
-    get: (href, data?, options?) => Inertia.get(href, data, mergeOptions(options)),
-    post: (href, data?, options?) => Inertia.post(href, data, mergeOptions(options)),
-    put: (href, data?, options?) => Inertia.put(href, data, mergeOptions(options)),
-    patch: (href, data?, options?) => Inertia.patch(href, data, mergeOptions(options)),
-    delete: (href, options?) => Inertia.delete(href, mergeOptions(options)),
+    visit: (href, options?) => Inertia.visit(href, extendInertiaEventCallbacks(node, options)),
+    reload: (options?) => Inertia.reload(extendInertiaEventCallbacks(node, options)),
+    get: (href, data?, options?) => Inertia.get(href, data, extendInertiaEventCallbacks(node, options)),
+    post: (href, data?, options?) => Inertia.post(href, data, extendInertiaEventCallbacks(node, options)),
+    put: (href, data?, options?) => Inertia.put(href, data, extendInertiaEventCallbacks(node, options)),
+    patch: (href, data?, options?) => Inertia.patch(href, data, extendInertiaEventCallbacks(node, options)),
+    delete: (href, options?) => Inertia.delete(href, extendInertiaEventCallbacks(node, options)),
   };
 }